Function rrise::communication::init

pub fn init(settings: AkCommSettings) -> Result<(), AkResult>

Initializes the communication module. When this is called, and [sound_engine::render_audio] is called periodically, you may use the authoring tool to connect to the sound engine.

Warning This function must be called after the sound engine and memory manager have been properly initialized.

Remark The AkCommSettings structure should be initialized with AkCommSettings::default. You can then change some of the parameters before calling this function.

Return

AK_Success if initialization was successful.

AK_InvalidParameter if one of the settings is invalid.

AK_InsufficientMemory if the specified pool size is too small for initialization.

AK_Fail for other errors.