Enum rrise::AkResult

#[repr(i32)]
pub enum AkResult {

    AK_NotImplemented,
    AK_Success,
    AK_Fail,
    AK_PartialSuccess,
    AK_NotCompatible,
    AK_AlreadyConnected,
    AK_InvalidFile,
    AK_AudioFileHeaderTooLarge,
    AK_MaxReached,
    AK_InvalidID,
    AK_IDNotFound,
    AK_InvalidInstanceID,
    AK_NoMoreData,
    AK_InvalidStateGroup,
    AK_ChildAlreadyHasAParent,
    AK_InvalidLanguage,
    AK_CannotAddItseflAsAChild,
    AK_InvalidParameter,
    AK_ElementAlreadyInList,
    AK_PathNotFound,
    AK_PathNoVertices,
    AK_PathNotRunning,
    AK_PathNotPaused,
    AK_PathNodeAlreadyInList,
    AK_PathNodeNotInList,
    AK_DataNeeded,
    AK_NoDataNeeded,
    AK_DataReady,
    AK_NoDataReady,
    AK_InsufficientMemory,
    AK_Cancelled,
    AK_UnknownBankID,
    AK_BankReadError,
    AK_InvalidSwitchType,
    AK_FormatNotReady,
    AK_WrongBankVersion,
    AK_FileNotFound,
    AK_DeviceNotReady,
    AK_BankAlreadyLoaded,
    AK_RenderedFX,
    AK_ProcessNeeded,
    AK_ProcessDone,
    AK_MemManagerNotInitialized,
    AK_StreamMgrNotInitialized,
    AK_SSEInstructionsNotSupported,
    AK_Busy,
    AK_UnsupportedChannelConfig,
    AK_PluginMediaNotAvailable,
    AK_MustBeVirtualized,
    AK_CommandTooLarge,
    AK_RejectedByFilter,
    AK_InvalidCustomPlatformName,
    AK_DLLCannotLoad,
    AK_DLLPathNotFound,
    AK_NoJavaVM,
    AK_OpenSLError,
    AK_PluginNotRegistered,
    AK_DataAlignmentError,
    AK_DeviceNotCompatible,
    AK_DuplicateUniqueID,
    AK_InitBankNotLoaded,
    AK_DeviceNotFound,
    AK_PlayingIDNotFound,
    AK_InvalidFloatValue,
    AK_FileFormatMismatch,
    AK_NoDistinctListener,
    AK_ACP_Error,
    AK_ResourceInUse,

}

Standard function call result.

Variants

AK_NotImplemented

This feature is not implemented.

AK_Success

The operation was successful.

AK_Fail

The operation failed.

AK_PartialSuccess

The operation succeeded partially.

AK_NotCompatible

Incompatible formats

AK_AlreadyConnected

The stream is already connected to another node.

AK_InvalidFile

An unexpected value causes the file to be invalid.

AK_AudioFileHeaderTooLarge

The file header is too large.

AK_MaxReached

The maximum was reached.

AK_InvalidID

The ID is invalid.

AK_IDNotFound

The ID was not found.

AK_InvalidInstanceID

The InstanceID is invalid.

AK_NoMoreData

No more data is available from the source.

AK_InvalidStateGroup

The StateGroup is not a valid channel.

AK_ChildAlreadyHasAParent

The child already has a parent.

AK_InvalidLanguage

The language is invalid (applies to the Low-Level I/O).

AK_CannotAddItseflAsAChild

It is not possible to add itself as its own child.

AK_InvalidParameter

Something is not within bounds.

AK_ElementAlreadyInList

The item could not be added because it was already in the list.

AK_PathNotFound

This path is not known.

AK_PathNoVertices

Stuff in vertices before trying to start it

AK_PathNotRunning

Only a running path can be paused.

AK_PathNotPaused

Only a paused path can be resumed.

AK_PathNodeAlreadyInList

This path is already there.

AK_PathNodeNotInList

This path is not there.

AK_DataNeeded

The consumer needs more.

AK_NoDataNeeded

The consumer does not need more.

AK_DataReady

The provider has available data.

AK_NoDataReady

The provider does not have available data.

AK_InsufficientMemory

Memory error.

AK_Cancelled

The requested action was cancelled (not an error).

AK_UnknownBankID

Trying to load a bank using an ID which is not defined.

AK_BankReadError

Error while reading a bank.

AK_InvalidSwitchType

Invalid switch type (used with the switch container)

AK_FormatNotReady

Source format not known yet.

AK_WrongBankVersion

The bank version is not compatible with the current bank reader.

AK_FileNotFound

File not found.

AK_DeviceNotReady

Specified ID doesn’t match a valid hardware device: either the device doesn’t exist or is disabled.

AK_BankAlreadyLoaded

The bank load failed because the bank is already loaded.

AK_RenderedFX

The effect on the node is rendered.

AK_ProcessNeeded

A routine needs to be executed on some CPU.

AK_ProcessDone

The executed routine has finished its execution.

AK_MemManagerNotInitialized

The memory manager should have been initialized at this point.

AK_StreamMgrNotInitialized

The stream manager should have been initialized at this point.

AK_SSEInstructionsNotSupported

The machine does not support SSE instructions (required on PC).

AK_Busy

The system is busy and could not process the request.

AK_UnsupportedChannelConfig

Channel configuration is not supported in the current execution context.

AK_PluginMediaNotAvailable

Plugin media is not available for effect.

AK_MustBeVirtualized

Sound was Not Allowed to play.

AK_CommandTooLarge

SDK command is too large to fit in the command queue.

AK_RejectedByFilter

A play request was rejected due to the MIDI filter parameters.

AK_InvalidCustomPlatformName

Detecting incompatibility between Custom platform of banks and custom platform of connected application

AK_DLLCannotLoad

Plugin DLL could not be loaded, either because it is not found or one dependency is missing.

AK_DLLPathNotFound

Plugin DLL search path could not be found.

AK_NoJavaVM

No Java VM provided in AkInitSettings.

AK_OpenSLError

OpenSL returned an error. Check error log for more details.

AK_PluginNotRegistered

Plugin is not registered. Make sure to implement a AK::PluginRegistration class for it and use AK_STATIC_LINK_PLUGIN in the game binary.

AK_DataAlignmentError

A pointer to audio data was not aligned to the platform’s required alignment (check AkTypes.h in the platform-specific folder)

AK_DeviceNotCompatible

Incompatible Audio device.

AK_DuplicateUniqueID

Two Wwise objects share the same ID.

AK_InitBankNotLoaded

The Init bank was not loaded yet, the sound engine isn’t completely ready yet.

AK_DeviceNotFound

The specified device ID does not match with any of the output devices that the sound engine is currently using.

AK_PlayingIDNotFound

Calling a function with a playing ID that is not known.

AK_InvalidFloatValue

One parameter has a invalid float value such as NaN, INF or FLT_MAX.

AK_FileFormatMismatch

Media file format unexpected

AK_NoDistinctListener

No distinct listener provided for AddOutput

AK_ACP_Error

Generic XMA decoder error.

AK_ResourceInUse

Resource is in use and cannot be released.