Module rrise::settings

Structs

AkCommSettings

Platform-independent initialization settings of communication module between the Wwise sound engine and authoring tool. \sa

AkDeviceSettings

High-level IO devices initialization settings. \sa

AkInitSettings

Platform-independent initialization settings of the sound engine See also

AkMemSettings

Initialization settings for the default implementation of the Memory Manager. For more details, see \ref memorymanager_init. \sa AK::MemoryMgr

AkPlatformInitSettings

\cond !(Web) Platform specific initialization settings \sa AK::SoundEngine::Init \sa AK::SoundEngine::GetDefaultPlatformInitSettings

AkStreamMgrSettings

Stream Manager initialization settings. \sa