Module rrise::sound_engine

Structs

PostEvent

Helper to post events to the sound engine.

Functions

add_default_listener

Add a single listener to the default set of listeners. Upon registration, all game objects reference the default listener set, until a call to add_listener, remove_listener, set_listeners or [set_game_object_output_bus_volume] is made on that game object.

add_listener

Add a single listener to a game object’s set of associated listeners.

init

Initialize the sound engine.

is_initialized

Query whether or not the sound engine has been successfully initialized.

load_bank_by_name

Load a bank synchronously (by Unicode string).

register_game_obj

Registers a game object.

register_named_game_obj

Registers a game object.

remove_default_listener

Remove a single listener from the default set of listeners. Upon registration, all game objects reference the default listener set, until a call to add_listener, remove_listener, set_listeners or [set_game_object_output_bus_volume] is made on that game object.

remove_listener

Remove a single listener from a game object’s set of active listeners.

render_audio

Processes all commands in the sound engine’s command queue.

set_default_listeners

Sets the default set of associated listeners for game objects that have not explicitly overridden their listener sets. Upon registration, all game objects reference the default listener set, until a call to add_listener, remove_listener, set_listeners or [set_game_object_output_bus_volume] is made on that game object.

set_listeners

Sets a game object’s associated listeners.

set_position

Sets the position of a game object.

stop_all

Stops the current content playing associated to the specified game object ID.

term

Terminates the sound engine.

unregister_all_game_obj

Unregister all game objects, or all game objects with a particular matching set of property flags.

unregister_game_obj

Unregisters a game object.