Function rrise::stream_mgr::set_current_language

pub fn set_current_language<T: AsRef<str>>(
    language_name: T
) -> Result<(), AkResult>

Set the current language once and only once, here. The language name is stored in a static buffer inside the Stream Manager. In order to resolve localized (language-specific) file location, the stream manger will query this string. It may use it to construct a file path (for e.g. SDK/samples/SoundEngine/Common/AkFileLocationBase.cpp), or to find a language-specific file within a look-up table (for e.g. SDK/samples/SoundEngine/Common/AkFilePackageLUT.cpp).

Pass a string, without a trailing slash or backslash. Empty strings are accepted.

You may register for language changes (see [register_to_language_change_notification]).

After changing the current language, all observers are notified.

Return AK_Success if successful (if language string has less than AK_MAX_LANGUAGE_NAME_SIZE characters). AK_Fail otherwise.

Warning Not multithread safe.

See also

[current_language]

[add_language_change_observer]