Struct rrise::AkCallbackType

#[repr(transparent)]
pub struct AkCallbackType(pub c_int);

Type of callback. Used as a bitfield in methods AK::SoundEngine::PostEvent() and AK::SoundEngine::DynamicSequence::Open().

Tuple Fields
0: c_int
Implementations

impl AkCallbackType

pub const AK_EndOfEvent: AkCallbackType = _

Callback triggered when reaching the end of an event. AkCallbackInfo can be cast to AkEventCallbackInfo.

impl AkCallbackType

pub const AK_EndOfDynamicSequenceItem: AkCallbackType = _

Callback triggered when reaching the end of a dynamic sequence item. AkCallbackInfo can be cast to AkDynamicSequenceItemCallbackInfo.

impl AkCallbackType

pub const AK_Marker: AkCallbackType = _

Callback triggered when encountering a marker during playback. AkCallbackInfo can be cast to AkMarkerCallbackInfo.

impl AkCallbackType

pub const AK_Duration: AkCallbackType = _

Callback triggered when the duration of the sound is known by the sound engine. AkCallbackInfo can be cast to AkDurationCallbackInfo.

impl AkCallbackType

pub const AK_SpeakerVolumeMatrix: AkCallbackType = _

Callback triggered at each frame, letting the client modify the speaker volume matrix. AkCallbackInfo can be cast to AkSpeakerVolumeMatrixCallbackInfo.

impl AkCallbackType

pub const AK_Starvation: AkCallbackType = _

Callback triggered when playback skips a frame due to stream starvation. AkCallbackInfo can be cast to AkEventCallbackInfo.

impl AkCallbackType

pub const AK_MusicPlaylistSelect: AkCallbackType = _

Callback triggered when music playlist container must select the next item to play. AkCallbackInfo can be cast to AkMusicPlaylistCallbackInfo.

impl AkCallbackType

pub const AK_MusicPlayStarted: AkCallbackType = _

Callback triggered when a “Play” or “Seek” command has been executed (“Seek” commands are issued from AK::SoundEngine::SeekOnEvent()). Applies to objects of the Interactive-Music Hierarchy only. AkCallbackInfo can be cast to AkEventCallbackInfo.

impl AkCallbackType

pub const AK_MusicSyncBeat: AkCallbackType = _

Enable notifications on Music Beat. AkCallbackInfo can be cast to AkMusicSyncCallbackInfo.

impl AkCallbackType

pub const AK_MusicSyncBar: AkCallbackType = _

Enable notifications on Music Bar. AkCallbackInfo can be cast to AkMusicSyncCallbackInfo.

impl AkCallbackType

pub const AK_MusicSyncEntry: AkCallbackType = _

Enable notifications on Music Entry Cue. AkCallbackInfo can be cast to AkMusicSyncCallbackInfo.

impl AkCallbackType

pub const AK_MusicSyncExit: AkCallbackType = _

Enable notifications on Music Exit Cue. AkCallbackInfo can be cast to AkMusicSyncCallbackInfo.

impl AkCallbackType

pub const AK_MusicSyncGrid: AkCallbackType = _

Enable notifications on Music Grid. AkCallbackInfo can be cast to AkMusicSyncCallbackInfo.

impl AkCallbackType

pub const AK_MusicSyncUserCue: AkCallbackType = _

Enable notifications on Music Custom Cue. AkCallbackInfo can be cast to AkMusicSyncCallbackInfo.

impl AkCallbackType

pub const AK_MusicSyncPoint: AkCallbackType = _

Enable notifications on Music switch transition synchronization point. AkCallbackInfo can be cast to AkMusicSyncCallbackInfo.

impl AkCallbackType

pub const AK_MusicSyncAll: AkCallbackType = _

Use this flag if you want to receive all notifications concerning AK_MusicSync registration.

impl AkCallbackType

pub const AK_MIDIEvent: AkCallbackType = _

Enable notifications for MIDI events. AkCallbackInfo can be cast to AkMIDIEventCallbackInfo.

impl AkCallbackType

pub const AK_CallbackBits: AkCallbackType = _

Bitmask for all callback types.

impl AkCallbackType

pub const AK_EnableGetSourcePlayPosition: AkCallbackType = _

Enable play position information for use by AK::SoundEngine::GetSourcePlayPosition().

impl AkCallbackType

pub const AK_EnableGetMusicPlayPosition: AkCallbackType = _

Enable play position information of music objects, queried via AK::MusicEngine::GetPlayingSegmentInfo().

impl AkCallbackType

pub const AK_EnableGetSourceStreamBuffering: AkCallbackType = _

Enable stream buffering information for use by AK::SoundEngine::GetSourceStreamBuffering().

impl AkCallbackType

pub fn contains(self, flags: Self) -> bool

Checks whether this bitflag has at least one of the bits in flags set.